Add convert_file function
This just takes two paths, and determines whether to convert from or to BIP based on the extensions.
It's called convert_file because I plan on adding convert_folder later, which will take a single folder path, and I guess some settings like whether to work recursively, whether to convert to or from bip, whether to delete source files after conversion, perhaps a list of extensions or a pattern to search with, I don't know.