Use size instead of width, height
I always need these in a tuple, so why not return them as a tuple.
And path will never be a Path, so I'm making that just str.
Also I'm simplifying type hinting a little, so that things will fit on one line in the other file (doesn't exist yet as of this commit).