Make preview collection clearing more robust
Now on load we make an event (if it's not there already), and send that to the worker function.
This way, we can control the worker function from outside by using the event.
The worker function also sends this event to the queue on the main thread, which in turn sends it to the function that loads data into the preview, IF the event isn't set.
On clear and on close, the event is set, and then set to None, which doesn't remove it from the functions and queue items that still have a reference to it, but does allow us to use a new Event for future load calls.
Niklas came up with this approach, and I think it's working, but will have to test with something that doesn't load instantly (perhaps just add some sleeps).