Check if name in self._collection
This check is necessary because loading a preview, then popping it before it's done loading, is a valid use case.
Also I'm changing some lines in new and load because I think this is more readable (you don't have to go look around for the definition of `__contains__` and `__getitem__`).