Use batching for load timer
Apparently timers have to wait their turn, in the same queue that includes regular (very slow) preview loads.
So when our timer gets called, we should make the most of it, by taking a whole 0.1 seconds to load as many previews as we can.
Perhaps I can implement a tag_redraw next, now that we have a way to not make it happen for every single preview.