Do a tag redraw after every batch
If this call to the timer has processed at least one item, do a tag redraw.
If we did not break out of the loop, meaning the queue still wasn't empty after the 100ms, make the delay 0.0, because we want to get back to loading ASAP.
The tag redraw affects every region of every area of every screen/window of Blender, because I don't know of any way to find out which areas are relevant to redraw.
Maybe take context in load and use that later?
Would only work if you call load from the UI you're drawing in.
Anyway this is fast enough I think.