Use trick to force preview load
Turns out the reason they would normally take so long is because of the scheduling and redraw after each load, so this is actually very fast.
It even works if you call load from the draw loop.
At the moment I don't see a reason to make it optional, because this preload behavior is already how it works for BIP and PIL, so it would be odd to have different behavior for this use case, especially when that behavior is generally inferior.
I still think BIP has a place, because it allows you to load any resolution, from a background thread, without dependencies.
I still think PIL has a place because it allows you to load any resolution, from a background thread, as long as you have Pillow installed.
But if you have neither, then this library can still be of great use, now that we can force previews to load quickly.