Implement resizing and lazy loading
Now you can choose whether to lazy load, which affects both BIP / PIL behavior and the fallback.
Also you can set a max size and bigger previews will be scaled down to fit.
When lazy loading is on, file loading is done in the thread pool, then preview loading is done in batches on the timer (or for the fallback it just loads them one by one when they're first drawn which means it's always 128x128).
When lazy loading is off, file loading is done on the main thread, and preview loading done immediately after on the same main thread; the timer is unused (or for the fallback it forces them to be loaded using the image_size trick, which means it's always 256x256).