Add warning when PIL is not installed
When you create a new collection, it lets you know that BIP images are not scaled and regular images are either 128x128 or 256x256.
We'll have to confirm later whether non-square aspect ratios are even supported by Blender, and perhaps change max_size to an int if they're not.
Also I'm importing annotations, purely because utils imports them and I want to be consistent.