Don't make pool and such when eager loading
The pool, event, queue, and timer are not created when we're not using lazy load.
When not lazy loading, call load eager, which creates the preview right there and fills it in.
I renamed load file to load async and load preview to load queued, and removed checks for lazy load from them.
I added checks for lazy load to clear and close, so that they don't bother with event, queue, and timer, when they don't need to.