Assert pixel count (#6)
* Assert if pixel count does not match size

Size being width times height.

* Improve ValueError docstring

Just making it a bit more accurate.

* Make clear and close more safer

Now the internal collection is only cleared / closed *after* clearning the queue / stopping the timer.

* Swap the order of safe and regular methods

I wanted to do this earlier but forgot.

* Don't make pool and such when eager loading

The pool, event, queue, and timer are not created when we're not using lazy load.
When not lazy loading, call load eager, which creates the preview right there and fills it in.
I renamed load file to load async and load preview to load queued, and removed checks for lazy load from them.
I added checks for lazy load to clear and close, so that they don't bother with event, queue, and timer, when they don't need to.

* Use load fallback for non-image previews too

Still need to actually test with MOVIE, FONT, and BLEND. I'd like to know whether the image_size trick would work them them too.
Also maybe we should add force_reload? Blender's ImagePreviewCollection has this option for load, but I'm not sure what it's for.

* Use image_size trick for non-image previews too

Tested, it works the same way.

* Use Git LFS for .mp4, .blend, and .ttf

I'm going to add a button to show some of these.

* Add misc files to display

One of each non-image type.
I think we're allowed to include this font.

* Add button to show misc previews

One of each type, loaded with the fallback method.
The image_size trick works with them, but you don't see it in action because this preview uses lazy loading.