Don't import __future__.annotations needlessly
I was like "but it's consistent to do it everywhere".
Thinking about it though, it's kind of silly.
If I wanted to be truly consistent I'd have to do it for every project, and I couldn't bring myself to do that, even if everyone on the team was fine with it (which they wouldn't be, and rightfully so).
So basically, yeet.