Use pre-multiplied alpha for converter
Apparently Blender needs it this way.
This does affect the format, but since the library is so young, Niklas says we don't need to go to BIP2 yet.
I was worried I'd have to remove the BIP -> PNG code, but it seems Pillow can convert from pre-multiplied alpha to regular alpha without issue.