Use pre-multiplied alpha for loader
For BIP it's easy, a small adjustment to the scaling code was all that had to be done on this end.

If an image has no alpha channel, it can not be converted to RGBa; but must instead be converted to RGBA.
After that conversion there's no need to pre-multiply the alpha, since it's 255 anyway.

If the source image did have an alpha channel, it is converted to RGBa, if it wasn't already.