Adjust loader code for readability
Removed some newlines which I felt didn't belong.
Renamed count to length when talking about (width * height), because that's what I call it elsewhere in the code too, and count refers to the amount of images in a BIP file.
Updated an assert message that I forgot to update earlier.
Changed how the data dictionary is used, so now icon_size and icon_pixels are set on its creation, and only overwritten in if the icon needed resizing.