import Image if PIL was installed by other addon
If Image isn't already set, check whether 'PIL' is in sys.modules (this is necessary to import from it).
If that is the case, then import it; not in a try block because this should never fail, so if it were to happen I'd rather get a big error than a silent failure.