Tweaks to thread.py
Adjusted some comments, newlines, formatting.
We want to catch it if a queue item is somehow not a correct size tuple, so I'm just doing a generic except instead of excepting Empty.
Variables that end in `sig` now end in `_signal`.
`_read_thread` no longer takes `thread_index`, as it wasn't using it.
Swapped the order of the signal check and try block.
Removed manual unregister of timer, as returning None is enough.